# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Trading Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /params:
    get:
      operationId: getParams
      tags: [params]
      summary: Get AI parameters
      description: Returns current AI tuning parameters
      responses:
        "200":
          description: Current parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiParams"
    patch:
      operationId: updateParams
      tags: [params]
      summary: Update AI parameters
      description: Manually adjust AI parameters
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAiParams"
      responses:
        "200":
          description: Updated parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiParams"
...
components:
  schemas:
...
    AiParams:
      type: object
      properties:
        minScore:
          type: number
        riskPerTrade:
          type: number
        updatedAt:
          type: string
          format: date-time
      required:
        - minScore
        - riskPerTrade
        - updatedAt
    UpdateAiParams:
      type: object
      properties:
        minScore:
          type: number
        riskPerTrade:
          type: number
...